# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "GridWiseController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "OpenAI key not configured",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Readiness; confirms server configuration, not live provider quota",
        "tags": [
          "GridWise"
        ]
      }
    },
    "/optimize-energy": {
      "post": {
        "description": "OpenAI structured extraction → deterministic guardrails → linear program → independent schedule replay. No authentication required. All energy units are kWh, currency BDT. Windows include start and exclude end.",
        "operationId": "GridWiseController_optimize",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scenario_id": {
                    "type": "string"
                  },
                  "operator_notes": {
                    "minItems": 1,
                    "maxItems": 3,
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "hours": {
                    "minItems": 24,
                    "maxItems": 24,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "hour": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 23
                        },
                        "demand_kwh": {
                          "type": "number",
                          "minimum": 0
                        },
                        "solar_kwh": {
                          "type": "number",
                          "minimum": 0
                        },
                        "tariff_bdt_per_kwh": {
                          "type": "number",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "hour",
                        "demand_kwh",
                        "solar_kwh",
                        "tariff_bdt_per_kwh"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "battery": {
                    "type": "object",
                    "properties": {
                      "capacity_kwh": {
                        "type": "number",
                        "minimum": 0
                      },
                      "initial_energy_kwh": {
                        "type": "number",
                        "minimum": 0
                      },
                      "minimum_energy_kwh": {
                        "type": "number",
                        "minimum": 0
                      },
                      "max_charge_kwh_per_hour": {
                        "type": "number",
                        "minimum": 0
                      },
                      "max_discharge_kwh_per_hour": {
                        "type": "number",
                        "minimum": 0
                      }
                    },
                    "required": [
                      "capacity_kwh",
                      "initial_energy_kwh",
                      "minimum_energy_kwh",
                      "max_charge_kwh_per_hour",
                      "max_discharge_kwh_per_hour"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "scenario_id",
                  "operator_notes",
                  "hours",
                  "battery"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scenario_id": {
                      "type": "string"
                    },
                    "directive_interpretation": {
                      "minItems": 1,
                      "maxItems": 3,
                      "type": "array",
                      "items": {
                        "oneOf": [
                          {
                            "type": "object",
                            "properties": {
                              "note_index": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 2
                              },
                              "explanation": {
                                "type": "string",
                                "minLength": 1
                              },
                              "applies": {
                                "type": "boolean",
                                "enum": [
                                  true
                                ]
                              },
                              "directive_type": {
                                "type": "string",
                                "enum": [
                                  "solar_reduction"
                                ]
                              },
                              "structured_adjustment": {
                                "type": "object",
                                "properties": {
                                  "hours": {
                                    "minItems": 1,
                                    "maxItems": 24,
                                    "type": "array",
                                    "items": {
                                      "type": "integer",
                                      "minimum": 0,
                                      "maximum": 23
                                    }
                                  },
                                  "factor": {
                                    "type": "number",
                                    "minimum": 0,
                                    "maximum": 1
                                  }
                                },
                                "required": [
                                  "hours",
                                  "factor"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "required": [
                              "note_index",
                              "explanation",
                              "applies",
                              "directive_type",
                              "structured_adjustment"
                            ],
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "properties": {
                              "note_index": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 2
                              },
                              "explanation": {
                                "type": "string",
                                "minLength": 1
                              },
                              "applies": {
                                "type": "boolean",
                                "enum": [
                                  true
                                ]
                              },
                              "directive_type": {
                                "type": "string",
                                "enum": [
                                  "minimum_battery_reserve"
                                ]
                              },
                              "structured_adjustment": {
                                "type": "object",
                                "properties": {
                                  "hours": {
                                    "minItems": 1,
                                    "maxItems": 24,
                                    "type": "array",
                                    "items": {
                                      "type": "integer",
                                      "minimum": 0,
                                      "maximum": 23
                                    }
                                  },
                                  "minimum_energy_kwh": {
                                    "type": "number",
                                    "minimum": 0
                                  }
                                },
                                "required": [
                                  "hours",
                                  "minimum_energy_kwh"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "required": [
                              "note_index",
                              "explanation",
                              "applies",
                              "directive_type",
                              "structured_adjustment"
                            ],
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "properties": {
                              "note_index": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 2
                              },
                              "explanation": {
                                "type": "string",
                                "minLength": 1
                              },
                              "applies": {
                                "type": "boolean",
                                "enum": [
                                  true
                                ]
                              },
                              "directive_type": {
                                "type": "string",
                                "enum": [
                                  "no_charge_window"
                                ]
                              },
                              "structured_adjustment": {
                                "type": "object",
                                "properties": {
                                  "hours": {
                                    "minItems": 1,
                                    "maxItems": 24,
                                    "type": "array",
                                    "items": {
                                      "type": "integer",
                                      "minimum": 0,
                                      "maximum": 23
                                    }
                                  }
                                },
                                "required": [
                                  "hours"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "required": [
                              "note_index",
                              "explanation",
                              "applies",
                              "directive_type",
                              "structured_adjustment"
                            ],
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "properties": {
                              "note_index": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 2
                              },
                              "explanation": {
                                "type": "string",
                                "minLength": 1
                              },
                              "applies": {
                                "type": "boolean",
                                "enum": [
                                  true
                                ]
                              },
                              "directive_type": {
                                "type": "string",
                                "enum": [
                                  "no_discharge_window"
                                ]
                              },
                              "structured_adjustment": {
                                "type": "object",
                                "properties": {
                                  "hours": {
                                    "minItems": 1,
                                    "maxItems": 24,
                                    "type": "array",
                                    "items": {
                                      "type": "integer",
                                      "minimum": 0,
                                      "maximum": 23
                                    }
                                  }
                                },
                                "required": [
                                  "hours"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "required": [
                              "note_index",
                              "explanation",
                              "applies",
                              "directive_type",
                              "structured_adjustment"
                            ],
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "properties": {
                              "note_index": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 2
                              },
                              "explanation": {
                                "type": "string",
                                "minLength": 1
                              },
                              "applies": {
                                "type": "boolean",
                                "enum": [
                                  true
                                ]
                              },
                              "directive_type": {
                                "type": "string",
                                "enum": [
                                  "max_grid_window"
                                ]
                              },
                              "structured_adjustment": {
                                "type": "object",
                                "properties": {
                                  "hours": {
                                    "minItems": 1,
                                    "maxItems": 24,
                                    "type": "array",
                                    "items": {
                                      "type": "integer",
                                      "minimum": 0,
                                      "maximum": 23
                                    }
                                  },
                                  "max_grid_kwh": {
                                    "type": "number",
                                    "minimum": 0
                                  }
                                },
                                "required": [
                                  "hours",
                                  "max_grid_kwh"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "required": [
                              "note_index",
                              "explanation",
                              "applies",
                              "directive_type",
                              "structured_adjustment"
                            ],
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "properties": {
                              "note_index": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 2
                              },
                              "explanation": {
                                "type": "string",
                                "minLength": 1
                              },
                              "applies": {
                                "type": "boolean",
                                "enum": [
                                  false
                                ]
                              },
                              "directive_type": {
                                "type": "string",
                                "enum": [
                                  "no_op"
                                ]
                              },
                              "structured_adjustment": {
                                "type": "string",
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            },
                            "required": [
                              "note_index",
                              "explanation",
                              "applies",
                              "directive_type",
                              "structured_adjustment"
                            ],
                            "additionalProperties": false
                          }
                        ]
                      }
                    },
                    "hourly_plan": {
                      "minItems": 24,
                      "maxItems": 24,
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "hour": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 23
                          },
                          "grid_kwh": {
                            "type": "number",
                            "minimum": 0
                          },
                          "solar_used_kwh": {
                            "type": "number",
                            "minimum": 0
                          },
                          "battery_action": {
                            "type": "string",
                            "enum": [
                              "charge",
                              "discharge",
                              "idle"
                            ]
                          },
                          "battery_kwh": {
                            "type": "number",
                            "minimum": 0
                          },
                          "battery_energy_after_kwh": {
                            "type": "number",
                            "minimum": 0
                          }
                        },
                        "required": [
                          "hour",
                          "grid_kwh",
                          "solar_used_kwh",
                          "battery_action",
                          "battery_kwh",
                          "battery_energy_after_kwh"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "total_grid_kwh": {
                      "type": "number",
                      "minimum": 0
                    },
                    "total_cost_bdt": {
                      "type": "number",
                      "minimum": 0
                    },
                    "peak_grid_kwh": {
                      "type": "number",
                      "minimum": 0
                    },
                    "plan_summary": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "scenario_id",
                    "directive_interpretation",
                    "hourly_plan",
                    "total_grid_kwh",
                    "total_cost_bdt",
                    "peak_grid_kwh",
                    "plan_summary"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON or invalid request structure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid battery bounds or infeasible constraints",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Controlled model/provider/internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Interpret 1–3 operator notes and return a minimum-cost 24-hour schedule",
        "tags": [
          "GridWise"
        ]
      }
    }
  },
  "info": {
    "title": "GridWise Energy Optimization API",
    "description": "BUP CSE Fest 2026. LLM interpretation, deterministic validation and optimal energy scheduling. Secrets remain server-side.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {}
  }
}